Default format JSON, don't overwrite output files by default
Change default output format for downloading data to JSON. Also, do not
overwrite output file if it already exists unless force is applied.

+++ ActionDownload.py
@@ -18,10 +18,16 @@
 #
 #	Johannes Bauer <JohannesBauer@gmx.de>
 
+import os
+import sys
 from BaseAction import BaseAction
 
 class ActionDownload(BaseAction):
 	def run(self):
+		if os.path.exists(self._args.output) and not (self._args.force):
+			self._log.error("Refusing to overwrite output file %s", self._args.output)
+			sys.exit(1)
+
 		data = self._rc4dev.readout()
 		with open(self._args.output, "w") as f:
 			if self._args.format == "txt":
